+++ library/Imbo/EventListener/StatsAccess.php
@@ -11,7 +11,6 @@
 namespace Imbo\EventListener;
 
 use Imbo\EventManager\EventInterface,
-    Imbo\Http\Request\Request,
     Imbo\Exception\RuntimeException;
 
 /**