+++ library/Imbo/Image/Transformation/Desaturate.php
@@ -10,8 +10,7 @@
 
 namespace Imbo\Image\Transformation;
 
-use Imbo\Model\Image,
-    Imbo\Exception\TransformationException,
+use Imbo\Exception\TransformationException,
     Imbo\EventListener\ListenerInterface,
     Imbo\EventManager\EventInterface,
     ImagickException;